properties: add properties prefix as filter/append
Properties may contain multiple objects based on prefix, for example:

    a.b.c.name =
    a.b.c.id =
    x.y.z.name =
    x.y.z.color =

To read these formats a prefix should be used as a filter when reading and as
append when generating.

A new optional JavaPropsSchema::prefix() and JavaPropsSchema::withPrefix() are
added to control the behavior.

As each path component is an object, use of ObjectReader::withRootName has a
different meaning so schema property was used.

PR #100